pyproject.toml
[project]
name = "tBot"
version = "0.1.2"
description = "Telegram bot for my personal use"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot==22.0",
    "python-dotenv>=1.0.0",
]
